#901 Update covered modules, move utility tests to correct project
