+++ hartshorn-util/src/test/java/test/org/dockbox/hartshorn/util/TupleTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package test.org.dockbox.hartshorn;
+package test.org.dockbox.hartshorn.util;
 
 import org.dockbox.hartshorn.util.Tuple;
 import org.junit.jupiter.api.Assertions;
