+++ settings.gradle.kts
@@ -29,7 +29,6 @@ includeReconfigured(
         ":hartshorn-proxy",
         ":hartshorn-proxy:hartshorn-proxy-cglib",
         ":hartshorn-proxy:hartshorn-proxy-javassist",
-        ":hartshorn-cdi",
         ":hartshorn-reporting",
         ":hartshorn-hsl",
         ":hartshorn-config",
@@ -43,7 +42,6 @@ includeReconfigured(
         ":hartshorn-jpa:hartshorn-jpa-hibernate",
         ":hartshorn-web",
         ":hartshorn-web:hartshorn-web-jetty",
-        ":hartshorn-web:hartshorn-web-freemarker",
         ":hartshorn-mvc",
         ":hartshorn-mvc:hartshorn-mvc-freemarker",
         ":hartshorn-test-suite",
